_sidecar/pyproject.toml
[project]
name = "tessa-design-sidecar"
version = "1.0.0"
description = "Design studio Python toolkit — research, Figma, clustering, video prompts, moodboards"
requires-python = ">=3.12"
dependencies = [
    # PDF + document processing
    "pypdf>=4.0",
    "pdfplumber>=0.11",
    # Semantic search + embeddings
    "ollama>=0.3",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    # Data
    "polars>=0.20",
    # Image processing
    "pillow>=10.0",
    "httpx>=0.27",
    # Figma API
    "httpx>=0.27",
    # CLI + output
    "typer>=0.12",
    "rich>=13.0",
    # Export
    "markdown>=3.6",
    "jinja2>=3.1",
]
